example: Add four part tinycore example
Add example yaml files to run a tinycore vm.  The 4 files implement the
following steps:
 1. Create base pvc to hold imported tinycore disk
 2. Import tinycore disk
 3. Clone tinycore pvc to a new pvc to use with vm
 4. Launch vm

To work around an issue in kubernetes, it is recommended to wait 15
seconds after creating a pvc before moving on to the next step.
Otherwise a delay of up to 4 minutes is possible when starting the pod.

Signed-off-by: Adam Litke <alitke@redhat.com>